Refactor normalizers
* add docs around drift detection
* move modules into subfolders to make import tree clear
* embed diffnormalize in diff package, both are small
